tcvpigpiv 0.3
First version that loads hourly ERA5 data from NCAR's RDA.

The main changes are: Slight modifications to the vpigpiv_module.py To ensure work with Hourly or monthly files.

Removal of the loader in the module and creating a new python function era5_loader.py to have a dedicated function just for loading

climatology.py that creates (and saves) climatology of the fields using 1980-2020. Then can be used to calculate anomalies on the hourly calculations (didn't test on monthly)

add a new notebook example to show how to work with the monthly, hourly, and anomalies.

Create the list of dependencies and requirements so if install in a new python environment it automatically downloads all requirements